♻️ refactor: 카테고리 이동 로직 및 라우팅 개선
