+++ src/ui/categoryButton.ts
@@ -6,12 +6,10 @@ export function categoryButton(category: Category): string {
   const route = isActive ? '/' : `/${category.id}`
 
   return `
-    <a href="${route}">
-      <button 
-        class="relative px-4 py-2 group category-button after:absolute after:left-0 after:bottom-0 after:h-0.5 after:bg-gray-900 ${activeClass}"
-        data-category="${category.id}">
-        <span>${category.label}</span>
-      </button>
-    </a>
+    <button 
+      class="relative px-4 py-2 group category-button after:absolute after:left-0 after:bottom-0 after:h-0.5 after:bg-gray-900 ${activeClass}"
+      onclick="navigate('${route}', false)">
+      <span>${category.label}</span>
+    </button>
   `
 }
